+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+- Add `CallableTree::Node#internal?`
+- Add `CallableTree::Node#external?`
+- (Experimental) Change the callables (matcher, caller, terminator) specified in the builder style to receive the current node instance as the `_node_` keyword argument.
+
 ## [0.3.6] - 2022-03-29
 
 - (Experimental) Add `CallableTree::Node::Hooks::Matcher`.